+++ tests/components/ipma/test_config_flow.py
@@ -3,20 +3,13 @@
 from unittest.mock import Mock, patch
 
 from homeassistant.components.ipma import DOMAIN, config_flow
-from homeassistant.const import CONF_LATITUDE, CONF_LONGITUDE, CONF_MODE, CONF_NAME
+from homeassistant.const import CONF_LATITUDE, CONF_LONGITUDE, CONF_MODE
 from homeassistant.helpers import entity_registry as er
 from homeassistant.setup import async_setup_component
 
-from .test_weather import MockLocation
+from . import MockLocation
 
 from tests.common import MockConfigEntry, mock_registry
-
-ENTRY_CONFIG = {
-    CONF_NAME: "Home Town",
-    CONF_LATITUDE: "1",
-    CONF_LONGITUDE: "2",
-    CONF_MODE: "hourly",
-}
 
 
 async def test_show_config_form():
